@@ -11,44 +11,44 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.0.3",
-    "@angular/cdk": "^6.4.0",
-    "@angular/common": "^6.0.3",
-    "@angular/compiler": "^6.0.3",
-    "@angular/core": "^6.0.3",
-    "@angular/forms": "^6.0.3",
-    "@angular/http": "^6.0.3",
-    "@angular/material": "^6.4.0",
-    "@angular/material-moment-adapter": "^6.4.0",
-    "@angular/platform-browser": "^6.0.3",
-    "@angular/platform-browser-dynamic": "^6.0.3",
-    "@angular/router": "^6.0.3",
-    "core-js": "^2.5.4",
+    "@angular/animations": "^7.1.2",
+    "@angular/cdk": "^6.4.7",
+    "@angular/common": "^7.1.2",
+    "@angular/compiler": "^7.1.2",
+    "@angular/core": "^7.1.2",
+    "@angular/forms": "^7.1.2",
+    "@angular/http": "^7.1.2",
+    "@angular/material": "^6.4.7",
+    "@angular/material-moment-adapter": "^6.4.7",
+    "@angular/platform-browser": "^7.1.2",
+    "@angular/platform-browser-dynamic": "^7.1.2",
+    "@angular/router": "^7.1.2",
+    "core-js": "^2.6.0",
     "material-design-icons": "^3.0.1",
     "moment": "^2.22.2",
-    "rxjs": "^6.2.2",
-    "rxjs-compat": "^6.2.2",
+    "rxjs": "^6.3.3",
+    "rxjs-compat": "^6.3.3",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.8",
-    "@angular/cli": "~6.0.8",
-    "@angular/compiler-cli": "^6.0.3",
-    "@angular/language-service": "^6.0.3",
-    "@types/jasmine": "~2.8.6",
-    "@types/jasminewd2": "~2.0.3",
+    "@angular-devkit/build-angular": "^0.11.2",
+    "@angular/cli": "~7.1.2",
+    "@angular/compiler-cli": "^7.1.2",
+    "@angular/language-service": "^7.1.2",
+    "@types/jasmine": "^2.8.12",
+    "@types/jasminewd2": "^2.0.6",
     "@types/node": "~8.9.4",
     "codelyzer": "~4.2.1",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^2.0.4",
+    "karma": "^2.0.5",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.0",
+    "karma-coverage-istanbul-reporter": "^2.0.4",
     "karma-jasmine": "~1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.3.0",
+    "protractor": "^5.4.1",
     "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "~2.7.2"
+    "typescript": "^3.1.6"
   }
 }
