@@ -1,3 +1,5 @@
+
+import {tap} from 'rxjs/operators';
 import { Injectable } from '@angular/core';
 import { HTTP_INTERCEPTORS } from '@angular/common/http';
 import { Observable } from 'rxjs';
@@ -17,12 +19,12 @@ export class HTTPRequestsInterceptor implements HttpInterceptor {
     ) {}
 
     intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
-        return next.handle(req).do(() => {}, err => {
+        return next.handle(req).pipe(tap(() => {}, err => {
             switch(err.status) {
             case 401: case 403:
                 this.profileService.Exit();
             }
-        });
+        }));
     }
 }
 
